requirements.txt
discord.py
chat-exporter
pytz
python-dotenv
aiohttp